backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
PyMuPDF
python-multipart
ollama  # If you're using local mistral
